requirements.txt
django-redis
sentry-sdk
requests
django
responses